Версия 12.2 Нужно доделать массив и тостер

+++ app/src/main/java/com/example/playlistmaker/Track.kt
@@ -5,5 +5,5 @@ data class Track(
     val artistName: String, // Имя исполнителя
     val trackTimeMillis: Long, // Продолжительность трека
     val artworkUrl100: String, // Ссылка на изображение обложки
-    val trackId: Int
+    val trackId: Long
 )
